% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-background.R
\name{add.background}
\alias{add.background}
\title{Calculate background as the average of the lowest five percent of data from data set}
\usage{
add.background(dat, Analyte = "CH4")
}
\arguments{
\item{dat:}{Data table, with names obtained from GMAP data output as of 2018-Aug}
}
\value{
None, but has the side effect of adding column CH4 and attribute bg to dat
}
\description{
Calculate background as the average of the lowest five percent of data from data set
}
\examples{
dat = add.background(dat)
}
