% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-db-OTM33A.R
\name{as.db.OTM33A}
\alias{as.db.OTM33A}
\title{Melt data from read.OTM33A into a database format}
\usage{
as.db.OTM33A(dat)
}
\arguments{
\item{dat}{Data table}
}
\value{

}
\description{
Melt data from read.OTM33A into a database format
}
\examples{
db=as.db.OTM33A(dat)
}
