% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDQI.R
\name{calcDQI}
\alias{calcDQI}
\title{Caculate DQI statistics at the end of the batch.
DQI statistics are documented in Appendix D to Appendix F1 of OTM 33A, v 1.3, p 53 to 59}
\usage{
calcDQI(dat)
}
\arguments{
\item{dat:}{Data table, after running emissions estimate calcs}
}
\value{
list of statistics used to calculate DQIs
}
\description{
Caculate DQI statistics at the end of the batch.
DQI statistics are documented in Appendix D to Appendix F1 of OTM 33A, v 1.3, p 53 to 59
}
\examples{
y = evalDQI(dat)
}
