% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDQIdb.R
\name{calcDQIdb}
\alias{calcDQIdb}
\title{Caculate DQI statistics at the end of the batch.
DQI statistics are documented in Appendix D to Appendix F1 of OTM 33A, v 1.3, p 53 to 59
This has lots of hacks to remain consistent with original ORD code.}
\usage{
calcDQIdb(datatable)
}
\arguments{
\item{db:}{Data table, database format for numerical data, after running emissions estimate calcs for CH4}
}
\value{
list of statistics used to calculate DQIs
}
\description{
Caculate DQI statistics at the end of the batch.
DQI statistics are documented in Appendix D to Appendix F1 of OTM 33A, v 1.3, p 53 to 59
This has lots of hacks to remain consistent with original ORD code.
}
\examples{
y = calcDQIdb(dat)
}
