% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmissionsRate.R
\name{calcEmissionsRate}
\alias{calcEmissionsRate}
\title{Calculate emissions rate for Analyte from data.table}
\usage{
calcEmissionsRate(dat, Analyte = "CH4")
}
\arguments{
\item{dat:}{Data table, with names obtained from GMAP data output as of 2018-Aug}
}
\value{
data.table with attributes set containing emissions calculations
}
\description{
Calculate emissions rate for Analyte from data.table
}
\examples{
dat = calcEmissionsRate(dat,Analyte="CH4")
}
