% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLyProjection-bin.R
\name{calcLyProjection.bin}
\alias{calcLyProjection.bin}
\title{Calculate plume direction and spread using binned data and natural scale}
\usage{
calcLyProjection.bin(dat, wdfilt = 90, binwidth = 2, plot = TRUE)
}
\arguments{
\item{dat:}{Data table, with names obtained from GMAP data output as of 2018-Aug
wdfilt:  number in degrees, used to exclude data from fit that are within the filter}
}
\value{
object of class "lm" with plume fit,
has the side effect of adding columns thetabin, theta.filter, Ly to dat
}
\description{
Calculate plume direction and spread using binned data and natural scale
}
\examples{
calcLyProjection.bin(dat)
}
