% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLyProjection-ind.R
\name{calcLyProjection.ind}
\alias{calcLyProjection.ind}
\title{Calculate plume direction and spread along axis perpendicular to the wind using individual data}
\usage{
calcLyProjection.ind(dat, wdfilt = 90, plot = TRUE)
}
\arguments{
\item{dat}{Data table, with names obtained from GMAP data output as of 2018-Aug}
}
\value{
object of class "lm" with plume fit
}
\description{
Calculate plume direction and spread along axis perpendicular to the wind using individual data
}
\examples{
calcLyProjection.ind(dat)
}
