% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlume.R
\name{calcPlume}
\alias{calcPlume}
\title{Calculate plume characteristics from data.table}
\usage{
calcPlume(dat)
}
\arguments{
\item{dat:}{Data table, with names obtained from GMAP data output as of 2018-Aug}
}
\value{
data.table with attributes set containing plume calculations
}
\description{
Calculate plume characteristics from data.table
}
\examples{
dat = calcPlume(dat)
}
