% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPSG-bin.R
\name{calcPSG.bin}
\alias{calcPSG.bin}
\title{Calculate plume direction and spread using binned data and natural scale}
\usage{
calcPSG.bin(dat, binwidth = 2, wdfilt = 180, plot = TRUE)
}
\arguments{
\item{dat}{Data table, with names obtained from GMAP data output as of 2018-Aug}
}
\value{
object of class "lm" with plume fit, has the side effect of adding column theta to dat
}
\description{
Calculate plume direction and spread using binned data and natural scale
}
\examples{
calcPSG.bin(dat)
}
