% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPSG-ind.R
\name{calcPSG.ind}
\alias{calcPSG.ind}
\title{Calculate plume direction and spread using indiviudal data on the natural scale}
\usage{
calcPSG.ind(dat, plot = TRUE)
}
\arguments{
\item{dat}{Data table, with names obtained from GMAP data output as of 2018-Aug}
}
\value{
object of class "lm" from plume fit
}
\description{
Calculate plume direction and spread using indiviudal data on the natural scale
}
\examples{
calcPSG.ind(dat)
}
