% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast-db.R
\name{cast.db}
\alias{cast.db}
\title{cast db back to spreadsheet form for analysis}
\usage{
cast.db(db, analyte = "CH4")
}
\arguments{
\item{db:}{Data table in db style developed from as.db.OTM33A
analyte:  Data table in spreadsheet format for use in analysis routines}
}
\value{
dat:
}
\description{
cast db back to spreadsheet form for analysis
}
\examples{
dat <- cast.db(db,analyte="CH4")
}
