% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval-DQI.R
\name{evalDQI}
\alias{evalDQI}
\title{Caculate DQI values from post-emissions-calc data set and DQI statistics from calcDQI.
DQI values are documented in Appendix D to Appendix F1 of OTM 33A, v 1.3, p 53 to 59}
\usage{
evalDQI(x, stats)
}
\arguments{
\item{dat:}{Data table, after running emissions estimate calcs
stats:  Output from calcDQI}
}
\value{
list of statistics used to calculate DQIs
}
\description{
Caculate DQI values from post-emissions-calc data set and DQI statistics from calcDQI.
DQI values are documented in Appendix D to Appendix F1 of OTM 33A, v 1.3, p 53 to 59
}
\examples{
y = evalDQI(dat)
}
