% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-results.R
\name{fetch.results}
\alias{fetch.results}
\title{Fetch final results from data}
\usage{
fetch.results(dat)
}
\arguments{
\item{dat:}{Data table, with names obtained from GMAP data output as of 2018-Aug}
}
\value{
Vector of named results, suitable for storing in a data.table
}
\description{
Fetch final results from data
}
\examples{
fetch.results(dat)
}
