% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-OTM33A.R
\name{read.OTM33A}
\alias{read.OTM33A}
\title{Read data from text file and return data.table}
\usage{
read.OTM33A(file.name, numskip = 33)
}
\arguments{
\item{dat}{Data table, with names obtained from GMAP data output as of 2018-Aug}
}
\value{
data.table containing data contained in file.name, with headers and attributes read from the file
}
\description{
Read data from text file and return data.table
}
\examples{
read.OTM33A(file.name,numskip=33)
}
