% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-STR.R
\name{read.STR}
\alias{read.STR}
\title{Read data from CSv file and return data.table}
\usage{
read.STR(file.name)
}
\arguments{
\item{dat}{Data table, with names obtained from STR data output from OTM3A Appendicies}
}
\value{
data.table containing data contained in file.name, with headers and attributes read from the file
}
\description{
Read data from CSv file and return data.table
}
\examples{
read.STR(file.name)
}
