% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename-STR.R
\name{rename.STR}
\alias{rename.STR}
\title{Rename data columns from read.OTM33A for subsequent processing}
\usage{
rename.STR(dat)
}
\arguments{
\item{dat}{Data table}
}
\description{
Rename data columns from read.OTM33A for subsequent processing
}
\examples{
dat=rename.OTM33A(dat)
}
