% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotateWindDirection.R
\name{rotateWindDirection}
\alias{rotateWindDirection}
\title{Calculate wind direction and caclulate wd3, the rotated wind direction variable}
\usage{
rotateWindDirection(dat)
}
\arguments{
\item{dat}{Data table, with names obtained from GMAP data output as of 2018-Aug}
}
\value{
void Columns are added to dat data table
}
\description{
Calculate wind direction and caclulate wd3, the rotated wind direction variable
}
\examples{
rotateWindDirection(dat)
}
