% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-batch.R
\name{run.batch}
\alias{run.batch}
\title{run batch files for a data set}
\usage{
run.batch(dat, wdfilt = 60, binwidth = 10)
}
\arguments{
\item{dat:}{Data table, with names obtained from GMAP data output as of 2018-Aug}
}
\value{

}
\description{
run batch files for a data set
}
\examples{
run.batch(dat)
}
