% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-fits.R
\name{run.fits}
\alias{run.fits}
\title{run batch files for a data set}
\usage{
run.fits(dat, wdfilt = 60, binwidth = 10, min.n = 0)
}
\arguments{
\item{dat:}{Data table, with names obtained from GMAP data output as of 2018-Aug}
}
\value{

}
\description{
run batch files for a data set
}
\examples{
run.fits(dat)
}
