% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-preprocess.R
\name{run.preprocess}
\alias{run.preprocess}
\title{Preprocess OTM33A data set prior to Gaussian fits}
\usage{
run.preprocess(dat)
}
\arguments{
\item{dat:}{Data table, with names obtained from GMAP data output as of 2018-Aug}
}
\value{

}
\description{
Preprocess OTM33A data set prior to Gaussian fits
}
\examples{
dat <- run.preprocess(dat)
}
