% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd-wd-yam.R
\name{sd.wd.yam}
\alias{sd.wd.yam}
\title{Calculate standard deviation of wind direction using Yamartino Method}
\usage{
sd.wd.yam(wd)
}
\arguments{
\item{dat:}{Wind direction,}
}
\value{
Numeric, value of standard deviation
}
\description{
Calculate standard deviation of wind direction using Yamartino Method
}
\examples{
wd.sd <- sd.wd.yam(wd)
}
