% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-align.R
\name{time.align}
\alias{time.align}
\title{Align sonic anenometer and concentration measurements in time by offsetting concentration measurements by a fixed delay, in seconds}
\usage{
\method{time}{align}(dat, delay)
}
\arguments{
\item{dat:}{Data table, with names obtained from GMAP data output as of 2018-Aug
delay:  Numeric, number of seconds in the delay}
}
\value{
None, but has the side effect of adding column CH4.nta (not time aligned) to dat
 overwrites CH4 with delayed signal
}
\description{
Align sonic anenometer and concentration measurements in time by offsetting concentration measurements by a fixed delay, in seconds
}
\examples{
time.align(dat, 40) # 40 second time delay
}
