% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmissionsRate.R
\name{calcEmissionsRate}
\alias{calcEmissionsRate}
\title{Calculate emissions rates from data.table and gaussian peak value estimated from calcLyProjection}
\usage{
calcEmissionsRate(dat, gauss.peak)
}
\arguments{
\item{dat:}{Data table, with names obtained from GMAP data output as of 2018-Aug

gauss.peak:  Peak of Gaussian fit from calcLyProjection}
}
\value{
vector of length 3 containing information from emissions rate calculation
}
\description{
Calculate emissions rates from data.table and gaussian peak value estimated from calcLyProjection
}
\examples{
calcEmissionsRate(file.name,numskip=33)
}
