% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename-OTM33A.R
\name{rename.OTM33A}
\alias{rename.OTM33A}
\title{Rename data columns from read.OTM33A for subsequent processing}
\usage{
rename.OTM33A(dat)
}
\arguments{
\item{dat}{Data table}
}
\description{
Rename data columns from read.OTM33A for subsequent processing
}
\examples{
dat=rename.OTM33A(dat)
}
